+++ core/src/main/java/com/lychcs/koikoi/scoring/ScoringEvent.java
@@ -1,3 +1,10 @@
 package com.lychcs.koikoi.scoring;
 
-public record ScoringEvent(String sourceName, int addedChips, int addedMult, double xMult, int addedMon, int addedVoidDust) {}
+public record ScoringEvent(
+    String sourceName,
+    long addedChips,
+    long addedMult,
+    double xMult,
+    int addedMon,
+    int addedVoidDust
+) {}
